build

+++ include/hs_server.h
@@ -52,7 +52,7 @@ void hs_server_on_connection(struct HSServer *, int /* socket */, void * /* cont
 /**
  * Simple utility function to create address for the given port.
  */
-struct sockaddr_in hs_server_init_ipv4_address(int /* port */);
+struct sockaddr_in hs_server_init_ipv4_address(uint16_t /* port */);
 
 #endif
